Restructure | Config settings
- Moved LanguageManager & Config into the Data package
- Moved some Objects to Utils/Objects
- Moved some comments in the config around.
- Added config options for the distance based system
