+++ src/main/java/me/deadlight/ezchestshop/Data/Config.java
@@ -1,5 +1,6 @@
-package me.deadlight.ezchestshop;
+package me.deadlight.ezchestshop.Data;
 
+import me.deadlight.ezchestshop.EzChestShop;
 import org.bukkit.configuration.file.FileConfiguration;
 
 public class Config {
@@ -9,6 +10,8 @@ public class Config {
     public static String firstLine;
     public static String secondLine;
     public static int holodelay;
+    public static boolean holodistancing;
+    public static double holodistancing_distance;
 
 
     public static void loadConfig() {
@@ -21,6 +24,9 @@ public class Config {
         firstLine = config.getString("hologram.hologram-first-line");
         secondLine = config.getString("hologram.hologram-second-line").replace("%currency%", Config.currency);
         holodelay = config.getInt("hologram.hologram-disappearance-delay");
+        holodistancing = config.getBoolean("hologram.distance.toggled");
+        holodistancing_distance = config.getDouble("hologram.distance.range");
+
 
     }
 }
