+++ src/main/java/me/deadlight/ezchestshop/GUIs/LogsGUI.java
@@ -1,8 +1,8 @@
 package me.deadlight.ezchestshop.GUIs;
 
-import me.deadlight.ezchestshop.LanguageManager;
+import me.deadlight.ezchestshop.Data.LanguageManager;
 import me.deadlight.ezchestshop.Utils.LogType;
-import me.deadlight.ezchestshop.Utils.TransactionLogObject;
+import me.deadlight.ezchestshop.Utils.Objects.TransactionLogObject;
 import me.deadlight.ezchestshop.Utils.Utils;
 import dev.triumphteam.gui.guis.Gui;
 import dev.triumphteam.gui.guis.GuiItem;
